Fix calculated dimensions breaking no-std
